refactor: rename dto folder to dtos
