+++ src/lib.rs
@@ -1,13 +1,13 @@
 mod builder;
 mod client;
 mod credentials;
-mod dto;
+mod dtos;
 mod enviroment;
 mod session;
 
 pub use builder::IQOptionBuilder;
 pub use client::IQOptionClient;
 pub use credentials::Credentials;
-pub use dto::*;
+pub use dtos::*;
 pub use enviroment::Enviroment;
 pub use session::Session;
